# fix(js): improve the `@nx/js/typescript` plugin performance