… packages/js/src/plugins/typescript/plugin.ts